==> 3. bigmart_testdatacleaning.ipynb <==
### Import Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder
import joblib

import warnings
warnings.filterwarnings('ignore')

### Read CSV file

test_data = pd.read_csv('test.csv')

### Display Data

test_data.head()

### Replicate handling of missing values and label encoding as done for train data

test_data['Item_Weight'] = test_data.groupby('Item_Type')['Item_Weight'].transform(lambda x: x.fillna(x.mean()))

test_data['Outlet_Size'] = test_data.groupby(['Outlet_Establishment_Year', 'Outlet_Location_Type', 'Outlet_Type'])['Outlet_Size'].transform(lambda x: x.fillna(x.mode().iloc[0] if not x.mode().empty else 'Unknown'))

test_data['Item_Visibility'] = test_data.groupby('Item_Identifier')['Item_Visibility'].transform(lambda x: x.replace(0, x.mean()))

test_data['Item_Fat_Content'] = test_data['Item_Fat_Content'].replace({'Low Fat': 1, 'low fat': 1, 'LF': 1, 'Regular': 2, 'reg': 2  })
test_data['Outlet_Size'] = test_data['Outlet_Size'].replace({'Unknown': 0, 'Small': 1, 'Medium': 2, 'High': 3 })
test_data['Outlet_Location_Type'] = test_data['Outlet_Location_Type'].replace({'Tier 1': 1, 'Tier 2': 2, 'Tier 3': 3 })
test_data['Outlet_Type'] = test_data['Outlet_Type'].replace({'Grocery Store': 0, 'Supermarket Type1': 1, 'Supermarket Type2': 2, 'Supermarket Type3': 3 })

test_data['Item_Fat_Content'] = test_data['Item_Fat_Content'].astype(int)
test_data['Outlet_Size'] = test_data['Outlet_Size'].astype(int)
test_data['Outlet_Location_Type'] = test_data['Outlet_Location_Type'].astype(int)

label_encoder_item = LabelEncoder()
label_encoder_outlet = LabelEncoder()
label_encoder_type = LabelEncoder()

# Fit and transform the 'Item_Identifier' column
test_data['Item_Identifier'] = label_encoder_item.fit_transform(test_data['Item_Identifier'])

# Fit and transform the 'Outlet_Identifier' column
test_data['Outlet_Identifier'] = label_encoder_outlet.fit_transform(test_data['Outlet_Identifier'])

# Fit and transform the 'Outlet_Identifier' column
test_data['Item_Type'] = label_encoder_type.fit_transform(test_data['Item_Type'])

### Save the test_data file

test_data.to_csv('test_data.csv', index=False)

### Save the label encoder instances to files

joblib.dump(label_encoder_item, 'label_encoder_item.joblib')
joblib.dump(label_encoder_outlet, 'label_encoder_outlet.joblib')
joblib.dump(label_encoder_type, 'label_encoder_type.joblib')
